+++ CHANGES.rst
@@ -7,6 +7,13 @@ CHANGES
   Added documentation for getting a client_id and secret.
   Deprecated "client_secret_filename" in config.
 * Issue 82: Feature request: Option to add phone number when creating new contact
+* Updated dependencies.
+* Issue 89: Support XDG Spec, files located in the old locations is still used if
+  they exists but XDG location are preferred.
+  ex.
+    $XDG_CONFIG_HOME/goobookrc
+    $XDG_CACHE_HOME/goobook_cache
+    $XDG_DATA_HOME/goobook_auth.json
 
 3.4 2019-09-10
 --------------